requirements.txt
Flask>=2.0,<3
rembg>=2.0
Pillow>=9.0
Werkzeug>=2.0
numpy>=1.24
onnxruntime>=1.15
gunicorn>=20.1